ecommerce/requirements.txt
phidata==2.7.6
duckdb==1.1.3
apify-client==1.8.1
openai==1.59.3
